extract note helpers into separate lib
